# perf(common): Improve performance of `StringInput`