@alexander-akait 2022-05-16
Should we do the same for HTML/CSS? Because I think we can apply this too